# Backend/api/openapi.yaml
openapi: 3.0.3
info:
  title: Fitness Tracker API
  version: 1.0.0
  description: |
    Lean Go backend for the core fitness tracking loop: users, exercise library,
    workouts, workout exercises, workout sets, meals, and weight entries.
servers:
  - url: https://localhost:8080
...
paths:
  /v1/leaderboard:
    get:
      tags: [Leaderboard]
      summary: Get Leaderboard
      operationId: getLeaderboard
      security:
        - BearerAuth: []
      parameters:
        - name: period
          in: query
          required: false
          schema:
            type: string
            enum: [weekly, monthly, yearly, alltime]
            default: weekly
        - name: pillar
          in: query
          required: false
          schema:
            type: string
            enum: [all, training, nutrition, consistency]
            default: all
        - name: offset
          in: query
          required: false
          schema:
            type: integer
            default: 0
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Leaderboard entries
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeaderboardResponse"
...
  /v1/rag/query:
    post:
      tags: [RAG]
      summary: Query the book RAG system
      operationId: queryRAG
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RAGQueryRequest"
      responses:
        "200":
          description: Answer generated from the RAG knowledge base.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RAGQueryResponse"
        default:
          $ref: "#/components/responses/Error"
  /v1/rag/cache/stats:
    get:
      tags: [RAG]
      summary: Get RAG cache statistics
      operationId: getRAGCacheStats
      responses:
        "200":
          description: Current RAG query cache metrics.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RAGCacheStatsResponse"
        default:
          $ref: "#/components/responses/Error"
  /v1/rag/cache/clear:
    post:
      tags: [RAG]
      summary: Clear the RAG query cache
      description: Admin-only endpoint that clears cached RAG query results and query-engine instances.
      operationId: clearRAGCache
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Cache clear completed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RAGCacheClearResponse"
        default:
          $ref: "#/components/responses/Error"
  /v1/auth/register:
    post:
      tags: [Auth]
      summary: Register a new user and create a session
      operationId: register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: User created and session established.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthSessionResponse"
        default:
          $ref: "#/components/responses/Error"
  /v1/auth/login:
    post:
      tags: [Auth]
      summary: Login with password and optional 2FA code (session-based)
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login succeeded. Returns access and refresh tokens.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthSessionResponse"
        "202":
          description: Password accepted and 2FA is required.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthSessionResponse"
        default:
          $ref: "#/components/responses/Error"
  /v1/auth/refresh:
    post:
      tags: [Auth]
      summary: Refresh an existing session
      operationId: refreshToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshRequest"
      responses:
        "200":
          description: New access and refresh tokens.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RefreshTokenResponse"
        default:
          $ref: "#/components/responses/Error"
  /v1/auth/logout:
    post:
      tags: [Auth]
      summary: Logout of current or all sessions
      operationId: logout
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LogoutRequest"
      responses:
        "204":
          $ref: "#/components/responses/NoContent"
        default:
          $ref: "#/components/responses/Error"
  /v1/auth/sessions:
    get:
      tags: [Auth]
      summary: List active sessions for the user
      operationId: getSessions
      responses:
        "200":
          description: List of active sessions.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedSessions"
        default:
          $ref: "#/components/responses/Error"
  /v1/auth/sessions/{id}:
    delete:
      tags: [Auth]
      summary: Revoke a specific session
      operationId: deleteSession
      parameters:
        - $ref: "#/components/parameters/ID"
      responses:
        "204":
          $ref: "#/components/responses/NoContent"
        default:
          $ref: "#/components/responses/Error"
  /v1/auth/2fa/setup:
    post:
      tags: [Auth]
      summary: Start 2FA setup for the authenticated user
      operationId: setupTwoFactor
      responses:
        "201":
          description: Secret and otpauth URL created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TwoFactorSetupResponse"
        default:
          $ref: "#/components/responses/Error"
  /v1/auth/2fa/verify:
    post:
      tags: [Auth]
      summary: Verify 2FA setup with a TOTP code
      operationId: verifyTwoFactor
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TwoFactorCodeRequest"
      responses:
        "200":
          description: 2FA enabled and recovery codes generated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TwoFactorVerifyResponse"
        default:
          $ref: "#/components/responses/Error"
  /v1/auth/2fa/disable:
    post:
      tags: [Auth]
      summary: Disable 2FA for the authenticated user
      operationId: disableTwoFactor
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TwoFactorCodeRequest"
      responses:
        "204":
          $ref: "#/components/responses/NoContent"
        default:
          $ref: "#/components/responses/Error"
  /v1/auth/2fa/recover:
    post:
      tags: [Auth]
      summary: Login with a recovery code
      operationId: recoverTwoFactor
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TwoFactorRecoveryLoginRequest"
      responses:
        "200":
          description: Login succeeded with recovery code.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthSessionResponse"
        default:
          $ref: "#/components/responses/Error"
  /v1/users:
    get:
      tags: [Users]
      summary: List users
      operationId: listUsers
      parameters:
        - in: query
          name: email
          schema:
            type: string
          description: Exact email filter.
        - in: query
          name: name
          schema:
            type: string
          description: Case-sensitive partial name filter.
      responses:
        "200":
          description: Users in descending creation order.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedUsers"
        default:
          $ref: "#/components/responses/Error"
    post:
      tags: [Users]
      summary: Create user
      operationId: createUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserRequest"
      responses:
        "201":
          description: Created user.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          $ref: "#/components/responses/Error"
  /v1/users/{id}:
    get:
      tags: [Users]
      summary: Get user
      operationId: getUser
      parameters:
        - $ref: "#/components/parameters/ID"
      responses:
        "200":
          description: User detail.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
    patch:
      tags: [Users]
      summary: Update user
      operationId: updateUser
      parameters:
        - $ref: "#/components/parameters/ID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserRequest"
      responses:
        "200":
          description: Updated user.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
    delete:
      tags: [Users]
      summary: Delete user
      operationId: deleteUser
      parameters:
        - $ref: "#/components/parameters/ID"
      responses:
        "204":
          $ref: "#/components/responses/NoContent"
        "400":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
        "500":
          $ref: "#/components/responses/Error"
  /v1/users/{user_id}/records:
    get:
      tags: [Analytics]
      summary: Get user personal records
      operationId: getUserRecords
      parameters:
        - $ref: "#/components/parameters/UserID"
      responses:
        "200":
          description: List of personal records.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedPersonalRecords"
        "400":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
  /v1/users/{user_id}/workout-stats:
    get:
      tags: [Analytics]
      summary: Get user workout stats
      operationId: getUserWorkoutStats
      parameters:
        - $ref: "#/components/parameters/UserID"
      responses:
        "200":
          description: Workout statistics.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkoutStats"
        "400":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
  /v1/users/{user_id}/activity-calendar:
    get:
      tags: [Adherence]
      summary: Get user activity calendar
      operationId: getActivityCalendar
      parameters:
        - $ref: "#/components/parameters/UserID"
        - in: query
          name: start
          schema:
            type: string
            format: date
        - in: query
          name: end
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Map of dates to activity types.
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  type: array
                  items:
                    type: string
        "400":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
  /v1/users/{user_id}/streaks:
    get:
      tags: [Adherence]
      summary: Get user streaks and adherence summary
      operationId: getUserStreaks
      parameters:
        - $ref: "#/components/parameters/UserID"
        - in: query
          name: date
          schema:
            type: string
            format: date
          description: Reference date for streak calculation (defaults to today).
      responses:
        "200":
          description: Streaks and adherence summary.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StreaksResponse"
        "400":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
  /v1/chat:
    post:
      tags: [Chat]
      summary: Send a message to the AI coach
      operationId: chat
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatRequest"
      responses:
        "200":
          description: AI response message.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatResponse"
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
  /v1/chat/history:
    get:
      tags: [Chat]
      summary: Get chat conversation history
      operationId: getChatHistory
      security:
        - BearerAuth: []
      parameters:
        - name: conversation_id
          in: query
          schema:
            type: string
            format: uuid
          description: Optional. If provided, returns messages for this conversation. Otherwise returns a list of conversations.
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
          description: Page number for the conversation list when `conversation_id` is omitted.
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
          description: Maximum number of conversations to return per page when `conversation_id` is omitted.
      responses:
        "200":
          description: List of conversations or conversation detail with messages.
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/PaginatedConversations"
                  - $ref: "#/components/schemas/Conversation"
        "401":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
  /v1/chat/feedback:
    post:
      tags: [Chat]
      summary: Provide feedback on an AI message
      operationId: chatFeedback
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatFeedbackRequest"
      responses:
        "200":
          description: Feedback received.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        "401":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
  /v1/users/{user_id}/coach-summary:
    get:
      tags: [AI Coach]
      summary: Get AI-generated coach summary for a user
      operationId: getCoachSummary
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/UserID"
        - name: date
          in: query
          schema:
            type: string
            format: date
          description: Optional. Date for the summary (defaults to today).
      responses:
        "200":
          description: Coach summary data.
          content:
            application/json:
              schema:
                type: object
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
...
components:
...
  responses:
    Error:
      description: Error response.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
...
    RAGQueryRequest:
      type: object
      required: [query]
      properties:
        query:
          type: string
          example: What is Kamal and what does it do?
        mode:
          type: string
          enum: [default, dense, sparse, hybrid]
          example: hybrid
        exact_search:
          type: boolean
          example: false
        similarity_top_k:
          type: integer
          minimum: 1
          example: 8
        sparse_top_k:
          type: integer
          minimum: 1
          example: 8
        include_sources:
          type: boolean
          example: true
...
    RAGQueryResponse:
      type: object
      required: [answer, mode_used]
      properties:
        answer:
          type: string
          example: Kamal is a deployment tool that helps you build, ship, and run apps on your own servers.
        mode_used:
          type: string
          example: hybrid
        sources:
          type: array
          items:
            $ref: "#/components/schemas/RAGQuerySource"
...
    LeaderboardResponse:
      type: object
      properties:
        period:
          type: string
        pillar:
          type: string
        total:
          type: integer
        offset:
          type: integer
        limit:
          type: integer
        entries:
          type: array
          items:
            $ref: '#/components/schemas/LeaderboardEntry'
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT